refactored bankroll from int to float

+++ src/main/java/com/arthur/blackjack/analytics/Analytics.java
@@ -2,7 +2,7 @@ package com.arthur.blackjack.analytics;
 
 public interface Analytics {
     void createNewResultsSheet(int bet);
-    void recordNewRound(Integer round, Integer money);
+    void recordNewRound(Integer round, float money);
     void recordInitialBet(Integer round, float bet);
     void recordRoundResult(Integer round, RoundResult result);
     void evaluateFormulas();
