+++ src/main/java/com/arthur/blackjack/config/GameSettings.java
@@ -1,12 +1,12 @@
 package com.arthur.blackjack.config;
 
 public interface GameSettings {
-    int getBankroll();
+    float getBankroll();
     int getBetSize();
     int getMinChipSize();
     int getBetSpread();
     int getMaxRounds();
-    void setBankroll(int bankroll);
+    void setBankroll(float bankroll);
     void setBetSize(int betSize);
     void setMinChipSize(int minChipSize);
     void setBetSpread(int betSpread);
